src/main/kotlin/mpdev/springboot/aoc2023/solutions/day12/Day12.kt
package mpdev.springboot.aoc2023.solutions.day12

import mpdev.springboot.aoc2023.model.PuzzlePartSolution
import mpdev.springboot.aoc2023.solutions.PuzzleSolver
import org.springframework.stereotype.Component
import kotlin.system.measureNanoTime
import kotlin.system.measureTimeMillis

@Component
class Day12: PuzzleSolver() {

    final override fun setDay() {
        day = 12
    }

    init {
        setDay()
    }

    var result = 0L
    lateinit var xxxx: Xxxx

    override fun initSolver(): Pair<Long,String> {
        val elapsed = measureNanoTime {
            xxxx = Xxxx(inputData)
        }
        return Pair(elapsed/1000, "micro-sec")
    }

    override fun solvePart1(): PuzzlePartSolution {
        val elapsed = measureTimeMillis {
            //result =
        }
        return PuzzlePartSolution(1, result.toString(), elapsed, "milli-sec")
    }

    override fun solvePart2(): PuzzlePartSolution {
        val elapsed = measureTimeMillis {
            //result =
        }
        return PuzzlePartSolution(2, result.toString(), elapsed, "milli-sec")
    }
}